feat: 业务线数据作用域隔离 + 跨租户泄漏门禁 + 历史结论复用
数据隔离（把业务线从"顺手加的字段"变成贯穿写入/读取/隔离三端的维度）：

- 现象与 DTC 改 (business_line, name|code) 复合唯一键：两者都是跨线共享
  的词汇表（实测 26 个现象名 13 个跨线、U0155 两条线都有），原单列唯一
  键会把两条线塌成一行/一个节点，属性被后写入者覆盖
- Neo4j 建点改复合 MERGE 键 + 唯一约束 + 业务线索引；查询补 business_line
  谓词（候选根因、现象词表、现象匹配）；作用域贯通四个入口（工具、REST、
  webhook 自动分诊、评测），关联问题单时以单子归属为准
- 去重去掉硬编码 "ia" 默认值（含降级暴力扫描路径），Milvus expr 按需拼接
- GraphRAG 强制作用域：NL2Cypher 必须引用 $business_line 参数，缺失或写死
  字面量一律拒绝执行并重试，重试耗尽返回空（不退化执行）
- impact 链路：删掉 "ia" 默认值，_check_scope 原先收下参数却没用，补上真过滤
- 业务线枚举进配置（BUSINESS_LINES），不再硬编码

RLS 管道：ContextVar + 事务级 set_config 钩子（SET LOCAL 语义，避免连接池
串租户）+ scripts/enable_rls.py（含前置检查：超级用户/表属主会绕过 RLS，
拒绝执行而不是给个"看起来开了"的假象）。裸 psycopg2 连接统一走
pg_scope.apply_scope —— 不接作用域会在 RLS 下静默返回空，不报错。

结论从个人记忆搬到知识层（记忆归人、结论归项目）：

- ai_triage_results 加 user_id / business_line（作用域缺失时从根因编码
  前缀推导，兜住 customer 会话），存量留 NULL 不回填
- 新增 search_past_diagnoses：双维度（我的 / 本业务线的）+ 角色门禁
  （客户只能查本人）+ 两级可信度（已采纳优先，降级未复核须带标记）
- 停止把诊断结论写长期记忆（记忆通道只承载人与车的事实），
  删掉 feedback_loop 里 106 行无调用点的历史复用死代码

存量修复与门禁：

- scripts/repair_line_split.py 拆分已塌缩的现象/DTC（PG + Neo4j 双端，
  干跑默认、幂等、保留边上的 weight/is_core）
- 跨租户泄漏集成门禁覆盖三条边界（图谱候选、历史结论、RLS 未设作用域）
- 静态守卫：裸 psycopg2 连接必须调用 apply_scope，AST 扫描 + 豁免登记

同时包含工作区内既有的前端、认证/限流、评测与部署改动。

+++ scripts/backfill_dedup_index.py
@@ -33,7 +33,7 @@ async def _load_issues(days: int, business_line: str | None) -> list[dict]:
     sql = """
         SELECT id, issue_no, business_line, title, description
         FROM alm_issues
-        WHERE updated_at > NOW() - (:days || ' days')::interval
+        WHERE updated_at > NOW() - ((:days)::int || ' days')::interval
     """
     params: dict = {"days": days}
     if business_line:
